TicketBookingBackEnd/ticket-reservation-app/src/main/java/com/guvi/ticket_reservation_app/service/AuthenticationService.java
package com.guvi.ticket_reservation_app.service;

import com.guvi.ticket_reservation_app.model.UserAuthenticationDetails;
import com.guvi.ticket_reservation_app.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class AuthenticationService implements UserDetailsService {
    @Autowired
    private UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        return userRepository.findByName(username)
                .map(UserAuthenticationDetails::new)
                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
    }
}
